requirements.txt
psutil

# Core Backend & API
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
requests
python-multipart

# AI & Machine Learning
tensorflow>=2.15.0
numpy
scikit-learn
joblib

# Auth
python-jose[cryptography]
passlib[bcrypt]

# Network & Real-time
scapy
websockets
httpx
redis

# Visualization & Utils
urllib3
